+++ src/core/footer/Footer.scss.d.ts
@@ -1,6 +1,7 @@
 declare namespace FooterScssNamespace {
   export interface IFooterScss {
     footer: string;
+    scriptVersion: string;
     selectLocale: string;
     selectedLocale: string;
   }