+++ src/plugins/dataTypes/Computed/Computed.generator.ts
@@ -8,7 +8,9 @@ export const generate = (data: DTGenerationData): DTGenerateResult => {
 
 	data.existingRowData.forEach((row: DTGenerationExistingRowData) => {
 		placeholders[`ROW${row.colIndex+1}`] = row.data.display;
+		placeholders[`ROWDATA${row.colIndex+1}`] = row;
 	});
+	placeholders.ROWNUM = data.rowNum;
 
 	return {
 		display: utils.generalUtils.template(data.rowState.value, placeholders)