+++ src/plugins/dataTypes/Computed/Computed.tsx
@@ -1,58 +1,108 @@
 import * as React from 'react';
-import { DTExampleProps, DTHelpProps, DTOptionsProps } from '~types/dataTypes';
+import { DTExampleProps, DTHelpProps, DTMetadata, DTOptionsProps } from '~types/dataTypes';
+import CopyToClipboard from '~components/CopyToClipboard/CopyToClipboard';
+import sharedStyles from '../../../styles/shared.scss';
+import styles from './Computed.scss';
 
 export const initialState = {
 	value: ''
 };
 
-export const Example = ({ i18n }: DTExampleProps): JSX.Element => i18n.see_help_dialog;
-
-export const Options = ({ data, onUpdate }: DTOptionsProps): JSX.Element => (
-	<textarea onChange={(e): void => onUpdate({ value: e.target.value })} value={data.value} />
+const Copy = ({ content, message, tooltip }: any): JSX.Element => (
+	<span className={styles.copy}>
+		<CopyToClipboard
+			tooltip={tooltip}
+			content={content}
+			message={message}
+		/>
+	</span>
 );
 
-export const Help = ({ i18n }: DTHelpProps): JSX.Element => (
+export const Example = ({ coreI18n }: DTExampleProps): JSX.Element => (
+	<div className={sharedStyles.emptyCol}>{coreI18n.seeHelpDialog}</div>
+);
+
+export const Options = ({ data, onUpdate }: DTOptionsProps): JSX.Element => (
+	<textarea
+		onChange={(e): void => onUpdate({ value: e.target.value })}
+		value={data.value}
+		style={{ width: '100%' }}
+	/>
+);
+
+export const Help = ({ coreI18n, i18n }: DTHelpProps): JSX.Element => (
 	<>
 		<p>
-			{i18n.help_para1}
+			{i18n.DESC}
 		</p>
 
-		<p>
-			{i18n.help_para2}
-		</p>
+		<div className={styles.row}>
+			<div className={styles.col1}>
+				<label>{`{{ROW1}}`}, {`{{ROW2}}`}, ...</label>
+			</div>
+			<div className={styles.copyCol}>
+				<Copy content="{{ROW1}}" message={coreI18n.copiedToClipboard} tooltip={coreI18n.copyToClipboard}/>
+			</div>
+			<div className={styles.col2}>
+				{i18n.rowPlaceholder}
+			</div>
+		</div>
+		<div className={styles.row}>
+			<div className={styles.col1}>
+				<label>{`{{ROWNUM}}`}</label>
+			</div>
+			<div className={styles.copyCol}>
+				<Copy content="{{ROWNUM}}" message={coreI18n.copiedToClipboard} tooltip={coreI18n.copyToClipboard}/>
+			</div>
+			<div className={styles.col2}>
+				{i18n.rowNumPlaceholder}
+			</div>
+		</div>
+		<div className={styles.row}>
+			<div className={styles.col1}>
+				<label>{`{{ROWDATA1}}`}, ...</label>
+			</div>
+			<div className={styles.copyCol}>
+				<Copy content="{{ROWDATA1}}" message={coreI18n.copiedToClipboard} tooltip={coreI18n.copyToClipboard}/>
+			</div>
+			<div className={styles.col2}>
+				{i18n.additionalInfo}
+				<ul>
+					<li><label>{`{{ROWDATA1.colIndex}}`}</label> - {i18n.colIndex}</li>
+					<li><label>{`{{ROWDATA1.id}}`}</label> - {i18n.randomUniqueId}</li>
+					<li><label>{`{{ROWDATA1.dataType}}`}</label> - {i18n.dataType}</li>
+					<li><label>{`{{ROWDATA1.data}}`}</label> - {i18n.generateDataObj}</li>
+				</ul>
+
+				{i18n.objExpl}
+				<label>{`{{JSON.stringify(ROWDATA1)}}`}</label>
+				<Copy content="{{JSON.stringify(ROWDATA1)}}" message={coreI18n.copiedToClipboard} tooltip={coreI18n.copyToClipboard}/>
+			</div>
+		</div>
+
+		<p dangerouslySetInnerHTML={{ __html: i18n.compositeHelp3 }} />
 
 		<ul>
-			<li>{i18n.help_prop1}</li>
-			<li>{i18n.help_prop2}</li>
-			<li>{i18n.help_prop3}</li>
-			<li>{i18n.help_prop4}</li>
-		</ul>
-
-		<b>{i18n.example}</b>
-
-		<ul>
-			<li>{i18n.example1}</li>
+			<li dangerouslySetInnerHTML={{ __html: i18n.compositeHelp4 }} />
+			<li>{i18n.compositeHelp5}
+				<ul>
+					<li><b>{`{{ROW2-ROW1}}`}</b> - {i18n.compositeSubtraction}</li>
+					<li><b>{`{{ROW2*ROW1}}`}</b> - {i18n.compositeMultiplication}</li>
+					<li><b>{`{{ROW2/ROW1}}`}</b> - {i18n.compositeDivision}</li>
+				</ul>
+			</li>
+			<li>
+				{i18n.compositeHelp6}
+				<b>{`{{ROW1 % 2 ? "even" : ROW2}}`}</b>
+			</li>
 		</ul>
 	</>
 );
 
-
-// var _validate = function(rows) {
-// 	var visibleProblemRows = [];
-// 	var problemFields      = [];
-// 	for (var i=0; i<rows.length; i++) {
-// 		if ($("#dtOption_" + rows[i]).val() === "") {
-// 			var visibleRowNum = generator.getVisibleRowOrderByRowNum(rows[i]);
-// 			visibleProblemRows.push(visibleRowNum);
-// 			problemFields.push($("#option_" + rows[i]));
-// 		}
-// 	}
-//
-// 	var errors = [];
-// 	if (visibleProblemRows.length) {
-// 		errors.push({ els: problemFields, error: L.AlphaNumeric_incomplete_fields + " <b>" + visibleProblemRows.join(", ") + "</b>"});
-// 	}
-//
-// 	return errors;
-// };
-//
+export const getMetadata = (): DTMetadata => ({
+	sql: {
+		field: 'TEXT default NULL',
+		field_Oracle: 'BLOB default NULL',
+		field_MSSQL: 'VARCHAR(MAX) NULL'
+	}
+});